Move some service registration from `launcher` to `build-state` and `daemon-services`.
