+++ platforms/core-runtime/build-state/src/main/java/org/gradle/internal/buildprocess/BuildProcessScopeServices.java
@@ -17,9 +17,18 @@
 package org.gradle.internal.buildprocess;
 
 import org.gradle.initialization.ClassLoaderRegistry;
+import org.gradle.internal.buildevents.BuildLoggerFactory;
+import org.gradle.internal.logging.LoggingManagerInternal;
 import org.gradle.internal.service.DefaultServiceLocator;
 import org.gradle.internal.service.ServiceRegistration;
+import org.gradle.internal.service.ServiceRegistry;
+import org.gradle.internal.service.scopes.GradleUserHomeScopeServiceRegistry;
 import org.gradle.internal.service.scopes.PluginServiceRegistry;
+import org.gradle.launcher.exec.BuildExecutor;
+import org.gradle.tooling.internal.provider.BuildSessionLifecycleBuildActionExecuter;
+import org.gradle.tooling.internal.provider.SessionFailureReportingActionExecuter;
+import org.gradle.tooling.internal.provider.SetupLoggingActionExecutor;
+import org.gradle.tooling.internal.provider.StartParamsValidatingActionExecuter;
 
 import java.util.List;
 
@@ -30,5 +39,22 @@ public class BuildProcessScopeServices {
             registration.add(PluginServiceRegistry.class, pluginServiceRegistry);
             pluginServiceRegistry.registerGlobalServices(registration);
         }
+    }
+
+    BuildExecutor createBuildExecuter(
+        LoggingManagerInternal loggingManager,
+        BuildLoggerFactory buildLoggerFactory,
+        GradleUserHomeScopeServiceRegistry userHomeServiceRegistry,
+        ServiceRegistry globalServices
+    ) {
+        return new SetupLoggingActionExecutor(
+            loggingManager,
+            new SessionFailureReportingActionExecuter(
+                buildLoggerFactory,
+                new StartParamsValidatingActionExecuter(
+                    new BuildSessionLifecycleBuildActionExecuter(userHomeServiceRegistry, globalServices)
+                )
+            )
+        );
     }
 }
