+++ platforms/core-runtime/launcher/src/main/java/org/gradle/tooling/internal/provider/LauncherServices.java
@@ -21,7 +21,6 @@ import org.gradle.api.internal.changedetection.state.FileHasherStatistics;
 import org.gradle.api.internal.tasks.userinput.BuildScanUserInputHandler;
 import org.gradle.api.internal.tasks.userinput.DefaultBuildScanUserInputHandler;
 import org.gradle.api.internal.tasks.userinput.DefaultUserInputHandler;
-import org.gradle.api.internal.tasks.userinput.DefaultUserInputReader;
 import org.gradle.api.internal.tasks.userinput.NonInteractiveUserInputHandler;
 import org.gradle.api.internal.tasks.userinput.UserInputHandler;
 import org.gradle.api.internal.tasks.userinput.UserInputReader;
@@ -45,13 +44,11 @@ import org.gradle.internal.buildtree.BuildTreeModelControllerServices;
 import org.gradle.internal.buildtree.InitDeprecationLoggingActionExecutor;
 import org.gradle.internal.buildtree.InitProblems;
 import org.gradle.internal.buildtree.ProblemReportingBuildActionRunner;
-import org.gradle.internal.classpath.CachedClasspathTransformer;
 import org.gradle.internal.concurrent.ExecutorFactory;
 import org.gradle.internal.enterprise.core.GradleEnterprisePluginManager;
 import org.gradle.internal.event.ListenerManager;
 import org.gradle.internal.execution.WorkInputListeners;
 import org.gradle.internal.file.StatStatistics;
-import org.gradle.internal.logging.LoggingManagerInternal;
 import org.gradle.internal.logging.sink.OutputEventListenerManager;
 import org.gradle.internal.logging.text.StyledTextOutputFactory;
 import org.gradle.internal.nativeintegration.filesystem.FileSystem;
@@ -61,9 +58,7 @@ import org.gradle.internal.operations.BuildOperationRunner;
 import org.gradle.internal.operations.logging.LoggingBuildOperationProgressBroadcaster;
 import org.gradle.internal.operations.notify.BuildOperationNotificationValve;
 import org.gradle.internal.service.ServiceRegistration;
-import org.gradle.internal.service.ServiceRegistry;
 import org.gradle.internal.service.scopes.AbstractPluginServiceRegistry;
-import org.gradle.internal.service.scopes.GradleUserHomeScopeServiceRegistry;
 import org.gradle.internal.session.BuildSessionActionExecutor;
 import org.gradle.internal.snapshot.CaseSensitivity;
 import org.gradle.internal.snapshot.ValueSnapshotter;
@@ -74,7 +69,6 @@ import org.gradle.internal.watch.vfs.BuildLifecycleAwareVirtualFileSystem;
 import org.gradle.internal.watch.vfs.FileChangeListeners;
 import org.gradle.internal.work.WorkerLeaseService;
 import org.gradle.launcher.exec.BuildCompletionNotifyingBuildActionRunner;
-import org.gradle.launcher.exec.BuildExecutor;
 import org.gradle.launcher.exec.BuildOutcomeReportingBuildActionRunner;
 import org.gradle.launcher.exec.BuildTreeLifecycleBuildActionExecutor;
 import org.gradle.launcher.exec.ChainingBuildActionRunner;
@@ -85,13 +79,6 @@ import org.gradle.problems.buildtree.ProblemDiagnosticsFactory;
 import org.gradle.problems.buildtree.ProblemReporter;
 import org.gradle.problems.buildtree.ProblemStream;
 import org.gradle.tooling.internal.provider.continuous.ContinuousBuildActionExecutor;
-import org.gradle.tooling.internal.provider.serialization.ClassLoaderCache;
-import org.gradle.tooling.internal.provider.serialization.DaemonSidePayloadClassLoaderFactory;
-import org.gradle.tooling.internal.provider.serialization.DefaultPayloadClassLoaderRegistry;
-import org.gradle.tooling.internal.provider.serialization.ModelClassLoaderFactory;
-import org.gradle.tooling.internal.provider.serialization.PayloadClassLoaderFactory;
-import org.gradle.tooling.internal.provider.serialization.PayloadSerializer;
-import org.gradle.tooling.internal.provider.serialization.WellKnownClassLoaderRegistry;
 
 import java.util.List;
 
@@ -101,12 +88,8 @@ import static org.gradle.internal.snapshot.CaseSensitivity.CASE_SENSITIVE;
 public class LauncherServices extends AbstractPluginServiceRegistry {
     @Override
     public void registerGlobalServices(ServiceRegistration registration) {
+        registration.add(BuildActionRunner.class, ExecuteBuildActionRunner.class);
         registration.addProvider(new ToolingGlobalScopeServices());
-    }
-
-    @Override
-    public void registerGradleUserHomeServices(ServiceRegistration registration) {
-        registration.addProvider(new ToolingGradleUserHomeScopeServices());
     }
 
     @Override
@@ -120,47 +103,8 @@ public class LauncherServices extends AbstractPluginServiceRegistry {
     }
 
     static class ToolingGlobalScopeServices {
-        void configure(ServiceRegistration registration) {
-            registration.add(UserInputReader.class, DefaultUserInputReader.class);
-            registration.add(BuildActionRunner.class, ExecuteBuildActionRunner.class);
-            registration.add(ClassLoaderCache.class, ClassLoaderCache.class);
-        }
-
-        BuildExecutor createBuildExecuter(
-            LoggingManagerInternal loggingManager,
-            BuildLoggerFactory buildLoggerFactory,
-            GradleUserHomeScopeServiceRegistry userHomeServiceRegistry,
-            ServiceRegistry globalServices
-        ) {
-            // @formatter:off
-            return
-                new SetupLoggingActionExecutor(loggingManager,
-                new SessionFailureReportingActionExecuter(buildLoggerFactory,
-                new StartParamsValidatingActionExecuter(
-                new BuildSessionLifecycleBuildActionExecuter(userHomeServiceRegistry, globalServices
-                ))));
-            // @formatter:on
-        }
-
         BuildLoggerFactory createBuildLoggerFactory(StyledTextOutputFactory styledTextOutputFactory, WorkValidationWarningReporter workValidationWarningReporter) {
             return new BuildLoggerFactory(styledTextOutputFactory, workValidationWarningReporter, Time.clock(), null);
-        }
-    }
-
-    static class ToolingGradleUserHomeScopeServices {
-        PayloadClassLoaderFactory createClassLoaderFactory(CachedClasspathTransformer cachedClasspathTransformer) {
-            return new DaemonSidePayloadClassLoaderFactory(
-                new ModelClassLoaderFactory(),
-                cachedClasspathTransformer);
-        }
-
-        PayloadSerializer createPayloadSerializer(ClassLoaderCache classLoaderCache, PayloadClassLoaderFactory classLoaderFactory) {
-            return new PayloadSerializer(
-                new WellKnownClassLoaderRegistry(
-                    new DefaultPayloadClassLoaderRegistry(
-                        classLoaderCache,
-                        classLoaderFactory))
-            );
         }
     }
